requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.8.2
